Remove the need for the `goxlr-ipc` crate.
Using the `goxlr-ipc` creates a dependency on specific utility versions in order to parse requests / responses into a DaemonStatus object. While this is preferred it has the issue that if the IPC structure changes, this app needs to be updated.

Given that we only use it to pull the Host / Port from the utility (their location hasn't changed since 0.12.0, and is unlikely to change in the future), we can just iterate over the json Value's instead of parsing out the entire DaemonStatus.